pyproject.toml
[project]
name = "aoc-input-fetcher"
version = "0.1.0"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
